chore: refactor docs command
